+++ src/main/java/tk/kirlian/DuckShop/errors/CannotTradeException.java
@@ -1,0 +1,7 @@
+package tk.kirlian.DuckShop.errors;
+
+/**
+ * Thrown to indicate a trade could not be completed.
+ */
+public class CannotTradeException extends DuckShopException {
+}