+++ src/main/java/tk/kirlian/DuckShop/errors/InvalidChestException.java
@@ -1,0 +1,7 @@
+package tk.kirlian.DuckShop.errors;
+
+/**
+ * Thrown to indicate a block should be a chest, but isn't.
+ */
+public class InvalidChestException extends DuckShopException {
+}