+++ src/main/java/tk/kirlian/DuckShop/errors/InvalidSyntaxException.java
@@ -1,9 +1,9 @@
-package tk.kirlian.SignTraderWithDucks.errors;
+package tk.kirlian.DuckShop.errors;
 
 /**
  * Thrown when a string cannot be parsed, usually in a TradingSign.
  */
-public class InvalidSyntaxException extends SignTraderException {
+public class InvalidSyntaxException extends DuckShopException {
     public InvalidSyntaxException() {
         super();
     }