+++ src/main/java/tk/kirlian/SignTraderWithDucks/errors/CannotTradeException.java
@@ -1,7 +1,0 @@
-package tk.kirlian.SignTraderWithDucks.errors;
-
-/**
- * Thrown to indicate a trade could not be completed.
- */
-public class CannotTradeException extends SignTraderException {
-}