+++ src/main/java/tk/kirlian/SignTraderWithDucks/errors/ChestProtectionException.java
@@ -1,7 +1,0 @@
-package tk.kirlian.SignTraderWithDucks.errors;
-
-/**
- * Thrown to indicate a chest is protected and cannot be accessed.
- */
-public class ChestProtectionException extends SignTraderException {
-}