+++ src/main/java/tk/kirlian/SignTraderWithDucks/errors/InvalidChestException.java
@@ -1,7 +1,0 @@
-package tk.kirlian.SignTraderWithDucks.errors;
-
-/**
- * Thrown to indicate a block should be a chest, but isn't.
- */
-public class InvalidChestException extends SignTraderException {
-}